refactor: clean libft.h
